+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Changelog
 ### next
+  * tpm2_import: support specifying parent key with a context file,
+    --parent-key-handle/-H becomes --parent-key/-C
   * tpm2_nvwrite and tpm2_nvread: when -P is "index" -a is optional and defaults to
     the NV_INDEX value passed to -x.
   * Load TCTI's by SONAME, not raw .so file